requirements.txt
aiocron==1.8
aiofiles==0.7.0
aiohttp==3.7.4.post0
APScheduler==3.10.4
Async-OAuthlib==0.0.9
async-timeout==3.0.1
asyncio==3.4.3
attrs==24.3.0
bcrypt==4.2.1
blinker==1.6.2
certifi==2024.12.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
croniter==1.4.1
cryptography==43.0.3
discord.py==1.7.3
exceptiongroup==1.2.2
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
idna==3.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.1
paramiko==3.5.0
pkcs7==0.1.2
pycparser==2.22
pycryptodome==3.19.0
PyNaCl==1.5.0
PyP100==0.1.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.22
starkbank-ecdsa==2.2.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.0.1
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
